Allow to hide resources block content
This block can take a lot of space on screen and it's not always
desirable to have it displayed by default.
This patch wraps the block content into a <details> element so the block
can be shown or hidden by a click.
It also adds a site setting to configure the default state of the block